+++ TestSuite/pack.py
@@ -217,7 +217,7 @@ class Pack:
                 "commonfields": {"id": name, "version": -1},
                 "name": name,
                 "display": name,
-                "description": f"this is an integration {name}",
+                "description": description or f"this is an integration {name}",
                 "category": "category",
                 "script": {
                     "type": "python",
@@ -226,6 +226,7 @@ class Pack:
                     "commands": [],
                     "dockerimage": docker_image,
                 },
+                "configuration": [],
             }
         if image is None:
             with open(