+++ demisto_sdk/commands/common/hook_validations/graph_validator.py
@@ -49,6 +49,7 @@ class GraphValidator(BaseValidator):
             self.validate_toversion_fields(),
             self.is_file_using_unknown_content(),
             self.is_file_display_name_already_exists(),
+            self.validate_duplicate_ids(),
         )
         return all(is_valid)
 
@@ -57,6 +58,25 @@ class GraphValidator(BaseValidator):
         is_valid = []
         is_valid.append(self.are_core_pack_dependencies_valid())
         return all(is_valid)
+
+    @error_codes("GR105")
+    def validate_duplicate_ids(self):
+        is_valid = True
+        for content_item, duplicates in self.graph.validate_duplicate_ids(
+            self.file_paths
+        ):
+            for duplicate in duplicates:
+                error_message, error_code = Errors.duplicated_id(
+                    content_item.object_id, duplicate.path
+                )
+                if self.handle_error(
+                    error_message,
+                    error_code,
+                    file_path=content_item.path,
+                    drop_line=True,
+                ):
+                    is_valid = False
+        return is_valid
 
     @error_codes("PA124")
     def are_core_pack_dependencies_valid(self):
@@ -207,13 +227,28 @@ class GraphValidator(BaseValidator):
     @error_codes("GR103")
     def is_file_using_unknown_content(self):
         """Validates that there is no usage of unknown content items.
+        The validation runs twice:
+        1. Cases where a warning should be raised - if the using content item is a test playbook/test script,
+            or if the dependency is optional.
+        2. Cases where an error should be raised - the complementary case.
+        """
+        is_valid = [
+            self._find_unknown_content_uses(raises_error=False),
+            self._find_unknown_content_uses(raises_error=True),
+        ]
+        return all(is_valid)
+
+    def _find_unknown_content_uses(self, raises_error: bool) -> bool:
+        """Validates that there is no usage of unknown content items.
         Note: if self.file_paths is empty, the validation runs on all files - in this case, returns a warning.
-        otherwise, returns an error.
+        otherwise, returns an error iff raises_error is True.
         """
 
         is_valid = True
         content_item: ContentItem
-        for content_item in self.graph.get_unknown_content_uses(self.file_paths):
+        for content_item in self.graph.get_unknown_content_uses(
+            self.file_paths, raises_error=raises_error
+        ):
             unknown_content_names = [
                 relationship.content_item_to.object_id or relationship.content_item_to.name  # type: ignore
                 for relationship in content_item.uses
@@ -225,7 +260,7 @@ class GraphValidator(BaseValidator):
                 error_message,
                 error_code,
                 content_item.path,
-                warning=not bool(self.file_paths),
+                warning=not bool(self.file_paths) or not raises_error,
             ):
                 is_valid = False