+++ demisto_sdk/commands/content_graph/objects/base_content.py
@@ -115,11 +115,7 @@ class BaseContent(ABC, BaseModel, metaclass=BaseContentMetaclass):
         logger.debug(f"Loading content item from path: {path}")
         if path.is_dir() and path.parent.name == "Packs":  # if the path given is a pack
             return content_type_to_model[ContentType.PACK].from_orm(PackParser(path))
-        try:
-            content_item_parser = ContentItemParser.from_path(path)
-        except Exception as e:
-            logger.error(f"Could not parse content item from path: {path}. Error: {e}")
-            return None
+        content_item_parser = ContentItemParser.from_path(path)
         if not content_item_parser:
             # This is a workaround because `create-content-artifacts` still creates deprecated content items
             demisto_sdk.commands.content_graph.parsers.content_item.MARKETPLACE_MIN_VERSION = (
@@ -131,7 +127,9 @@ class BaseContent(ABC, BaseModel, metaclass=BaseContentMetaclass):
             )
 
         if not content_item_parser:  # if we still can't parse the content item
-            logger.error(f"Could not parse content item from path: {path}")
+            logger.error(
+                f"Invalid content path provided: {str(path)}. Please provide a valid content item or pack path."
+            )
             return None
 
         model = content_type_to_model.get(content_item_parser.content_type)