+++ demisto_sdk/commands/format/update_description.py
@@ -1,7 +1,6 @@
+import logging
 import re
 from typing import Tuple
-
-import click
 
 from demisto_sdk.commands.common.constants import BETA_INTEGRATION_DISCLAIMER
 from demisto_sdk.commands.common.tools import find_type
@@ -13,6 +12,8 @@ from demisto_sdk.commands.format.format_constants import (
 from demisto_sdk.commands.format.update_generic import BaseUpdate
 
 CONTRIBUTOR_DETAILED_DESC = "Contributed Integration"
+
+logger = logging.getLogger("demisto-sdk")
 
 
 class DescriptionFormat(BaseUpdate):
@@ -29,7 +30,6 @@ class DescriptionFormat(BaseUpdate):
         path: str = "",
         from_version: str = "",
         no_validate: bool = False,
-        verbose: bool = False,
         update_docker: bool = False,
         **kwargs,
     ):
@@ -39,7 +39,6 @@ class DescriptionFormat(BaseUpdate):
             path=path,
             from_version=from_version,
             no_validate=no_validate,
-            verbose=verbose,
             **kwargs,
         )
         description_type = input.replace("_description.md", ".yml")
@@ -69,19 +68,16 @@ class DescriptionFormat(BaseUpdate):
 
     def save_md_to_destination_file(self):
         """Safely saves formatted YML data to destination file."""
-        if self.source_file != self.output_file and self.verbose:
-            click.secho(
-                f"Saving output description file to {self.output_file} \n", fg="white"
-            )
+        if self.source_file != self.output_file:
+            logger.debug(f"Saving output description file to {self.output_file} \n")
         with open(self.output_file, "w") as f:
             f.write(self.description_content)
         f.close()
 
     def run_format(self) -> int:
         try:
-            click.secho(
-                f"\n================= Updating file {self.source_file} ================= ",
-                fg="bright_blue",
+            logger.info(
+                f"\n[blue]================= Updating file {self.source_file} =================[/blue]"
             )
             self.remove_community_partner_details()
             if self.is_beta:
@@ -89,11 +85,9 @@ class DescriptionFormat(BaseUpdate):
             self.save_md_to_destination_file()
             return SUCCESS_RETURN_CODE
         except Exception as err:
-            if self.verbose:
-                click.secho(
-                    f"\nFailed to update file {self.source_file}. Error: {err}",
-                    fg="red",
-                )
+            logger.debug(
+                f"\n[red]Failed to update file {self.source_file}. Error: {err}[/red]"
+            )
             return ERROR_RETURN_CODE
 
     def format_file(self) -> Tuple[int, int]: