+++ CHANGELOG.md
@@ -1,21 +1,39 @@
 # Changelog
 
 ## Unreleased
-* Fixed an issue where a pack `serverMinVersion` is calculated by its content items to be the minimum fromVersion.
+* **Note: Demisto-SDK will soon stop supporting Python 3.8**
+* Fixed an issue where using **download** on non-unicode content, merging them into existing files caused an error.
+* Changed an internal setting to allow writing non-ascii content (unicode) using `YAMLHandler` and `JSONHandler`.
+* Fixed an issue where an error message in **unify** was unclear for invalid input.
+* Fixed an issue where running **validate** failed with **is_valid_integration_file_path_in_folder** on integrations that use API modules.
+* Fixed an issue where **validate** failed with **is_valid_integration_file_path_in_folder** on integrations that use the `MSAPIModule`.
+* Added **validate** check for the `modules` field in `pack_metadata.json` files.
+* Changed **lint** to skip deprecated content, unless when using the `-i` flag.
+* Fixed an issue where **update-release-notes** failed when a new *Parsing Rule* was added to a pack.
+* Refactored the logging framework. Demisto-SDK logs will now be written to `.demist_sdk_debug.log` under the content path (when detected) or the current directory.
+* Added `GR105` validation to **validate** command to check that no duplicate IDs are used.
+* Added support for API Modules imported in API modules in the **unify** command.
 * Added the **pre-commit** command, to run pre-commit hooks on XSOAR content.
 * Added the **run-unit-tests** command, to run unit tests of given content items inside their respective docker images.
 * Added support for filepath arguments in the **validate** and **format** commands.
 * Added pre-commit hooks for `validate`, `format`, `run-unit-tests` and `update-docker-image` commands.
+
+
+## 1.10.6
+* Fixed an issue where running **validate** with the `-g` flag would skip some validations for old-formatted (unified) integration/script files.
+* Deprecated integrations and scripts will not run anymore when providing the **--all-packs** to the **lint** command.
+* Fixed an issue where a pack `serverMinVersion` would be calculated by the minimal fromVersion of its content items.
+* Added the `--docker-image-target` flag to **lint** for testing native supported content with new images.
 
 ## 1.10.5
 * Fixed an issue where running **run-test-playbook** would not use the `verify` parameter correctly. @ajoga
 * Added a newline at the end of README files generated in **generate-docs**.
 * Added the value `3` (out of bounds) to the `onChangeRepAlg` and `reputationCalc` fields under the `IncidentType` and `GenericType` schemas. **validate** will allow using it now.
 * Fixed an issue where **doc-review** required dot suffixes in release notes describing new content.
+* Fixed an issue where **validate** failed on Feed Integrations after adding the new *Collect/Connect* section field.
 * Fixed an issue where using **postman-codegen** failed converting strings containing digits to kebab-case.
 * Fixed an issue where the ***error-code*** command could not parse List[str] parameter.
 * Updated validation *LO107* to support more section types in XSIAM layouts.
-* Deprecated integrations and scripts will not run anymore when providing the **--all-packs** to the **lint** command.
 
 ## 1.10.4
 * Added support for running **lint** in multiple native-docker images.